updated the score calculation logic so we always increase the score when the barriers are hit. so far only simple barrier affects the score

+++ src/main/java/Domain/HUD.java
@@ -34,7 +34,7 @@ public class HUD extends JPanel {
         livesLabel.setText("Lives: " + GameView.getLives());
     }
 
-    public void updateScore(int score) {
+    public void updateScore(long score) {
         scoreLabel.setText("Score: " + score);
     }
 }
